export_soc_helper: change end of simulation point
Simulate until the end of any production, instead of the end of excess
production. This should avoid the (potential) problem of charging being
disabled in the afternoon on the theory that it might leave the battery
too high the next morning. In practice, overnight usage will be higher
than min_discharge_power, and it is better to charge the battery as much
as possible in the afternoon.